chore(webpack-builder): move `src/shared/stub` to `src/stub`
