+++ packages/builder/webpack-builder/src/stub/builder.ts
@@ -1,9 +1,9 @@
 import _ from '@modern-js/utils/lodash';
 import assert from 'assert';
-import { createPrimaryBuilder } from '../../core/createBuilder';
-import { Hooks } from '../../core/createHook';
-import type { BuilderOptions, BuilderPlugin, Context } from '../../types';
-import { matchLoader, mergeBuilderOptions } from '../utils';
+import { createPrimaryBuilder } from '../core/createBuilder';
+import { Hooks } from '../core/createHook';
+import type { BuilderOptions, BuilderPlugin, Context } from '../types';
+import { matchLoader, mergeBuilderOptions } from '../shared';
 import { createStubContext } from './context';
 
 export interface StubBuilderOptions extends BuilderOptions {
