+++ packages/builder/webpack-builder/src/stub/index.ts
