+++ packages/builder/webpack-builder/tests/plugins/fallback.test.ts
@@ -1,7 +1,7 @@
 import { expect, describe, it } from 'vitest';
 import { PluginFallback } from '../../src/plugins/fallback';
 import { BuilderPlugin } from '../../src/types';
-import { createStubBuilder } from '../../src/shared/stub';
+import { createStubBuilder } from '../../src/stub';
 
 describe('plugins/fallback', () => {
   const testPlugin: BuilderPlugin = {
