+++ packages/builder/webpack-builder/tests/plugins/react.test.ts
@@ -2,7 +2,7 @@ import { expect, describe, it } from 'vitest';
 import { PluginReact } from '../../src/plugins/react';
 import { PluginBabel } from '../../src/plugins/babel';
 import { PluginTsLoader } from '../../src/plugins/tsLoader';
-import { createStubBuilder } from '../../src/shared/stub';
+import { createStubBuilder } from '../../src/stub';
 
 describe('plugins/react', () => {
   it('should work with babel-loader', async () => {
