+++ packages/builder/webpack-builder/tests/plugins/resolve.test.ts
@@ -1,6 +1,6 @@
 import { expect, describe, it } from 'vitest';
 import { PluginResolve } from '../../src/plugins/resolve';
-import { createStubBuilder } from '../../src/shared/stub';
+import { createStubBuilder } from '../../src/stub';
 
 describe('plugins/resolve', () => {
   it('should apply default extensions correctly', async () => {
